Group Auth

+++ backend/src/controllers/groups-controller.ts
@@ -1,23 +1,29 @@
 import { Request, Response } from 'express';
-import Group from '../models/group';
 import { validationResult } from 'express-validator';
 import { ObjectId } from "mongodb";
+import { CustomRequest } from "../middleware/auth";
+import Group from "../models/group";
+
 
 
 
 
 
 //Creates a group with the desired name and gives a unique ID
-export const createGroup = async (req: Request, res: Response) => {
+export const createGroup = async (req: CustomRequest, res: Response) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log("errors", errors);
         return res.status(400).json({ errors: errors.array() });
     }
     //Checks if there's a similar group
-    const dupeGroup = Group.findOne({ name: req.body })
-    if (!dupeGroup) {
+    const dupeGroup = await Group.findOne({ name: req.body.name })
+    if (dupeGroup !== null) {
         return res.status(409).json({ error: "A group with the same name already exists."  });
+    }
+    //Checks if UserId exists
+    if(!req.token?.userId){
+        return res.status(422).json({error: "UserId does not exist"});
     }
     else{
         //Creates a new group with name and id
@@ -26,6 +32,7 @@ export const createGroup = async (req: Request, res: Response) => {
 
             const group = new Group({
                 name: name,
+                userId: req.token.userId
             });
             await group.save();
             return res.status(201).json({group});
@@ -62,7 +69,7 @@ export const getSpecificGroup = async (req: Request, res: Response) => {
 }
 
 //Updates the name of an existing group
-export const updateName = async (req: Request, res: Response) =>{
+export const updateName = async (req: CustomRequest, res: Response) =>{
     const { newName }: { newName: string }= req.body;
     const groupID: string = req.params._id;
 
@@ -74,8 +81,21 @@ export const updateName = async (req: Request, res: Response) =>{
     }
     //Group existence error
     const existingGroup = await Group.findOne({name: newName})
-    if(existingGroup){
+    if(existingGroup!== null){
         return res.status(409).json({ error: "This group name is already taken"});
+    }
+    //Checks if UserId exists
+    if(!req.token?.userId){
+        return res.status(422).json({error: "UserId does not exist"});
+    }
+    const findGroupOwnerById = await Group.findOne({ userId: req.token.userId, _id: groupID });
+    //Makes sure user is owner of the item
+    if (!findGroupOwnerById){
+        return res.status(403).json({ error: 'The user id does not match.'});
+    }
+    //Token check
+    if (!req.token) {
+        return res.status(400).json({ message: 'Token missing' })
     }
     //Updates group
     try {
@@ -89,12 +109,24 @@ export const updateName = async (req: Request, res: Response) =>{
     catch (error: any) {
         res.status(500).json({ message: 'Server error', error: error.message });
     }
-
 }
 
 //Deletes group by ID
-export const deleteGroup = async (req: Request, res: Response) => {
+export const deleteGroup = async (req: CustomRequest, res: Response) => {
     const groupID: string = req.params._id;
+    //Checks if UserId exists
+    if(!req.token?.userId){
+        return res.status(422).json({error: "UserId does not exist"});
+    }
+    const findGroupOwnerById = await Group.findOne({ userId: req.token.userId, _id: groupID });
+    //Makes sure user is owner of the item
+    if (!findGroupOwnerById){
+        return res.status(403).json({ error: 'The user id does not match.'});
+    }
+    //Token check
+    if (!req.token) {
+        return res.status(400).json({ message: 'Token missing' })
+    }
     try {
         const result = await Group.deleteOne({_id: new ObjectId(groupID)});
         if(result.deletedCount === 0){
